[UpdateOnly] move each index kind under the appendable index it writes for
`numeric_index/update_only.rs` and its three siblings sat at the index-type
level, next to the enum over all three storage variants, although each writes
for the appendable variant alone. They now live at
`<index>/mutable_<index>/update_only/`, beside that variant's `read_only/`
counterpart, which is the same split for the same reason.

`mutable_text_index` is private, so the text kind is re-exported from
`full_text_index` for the dispatch enum to name.
