+++ lib/segment/src/index/field_index/map_index/mutable_map_index/mod.rs
@@ -7,6 +7,7 @@ pub(super) mod in_memory;
 mod lifecycle;
 pub mod read_only;
 mod read_ops;
+pub mod update_only;
 
 pub struct MutableMapIndex<N: MapIndexKey + ?Sized>
 where
