+++ lib/segment/src/index/field_index/numeric_index/mod.rs
@@ -10,7 +10,6 @@ mod query;
 mod read_only;
 mod read_ops;
 mod storage;
-pub mod update_only;
 mod value_indexer;
 
 use std::marker::PhantomData;
